refator: coupon 도메인에 couponFacade 추가, 내 쿠폰 사용 코드 개선
